requirements.txt
absl-py==0.11.0
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217285838/work
astunparse==1.6.3
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.9.3
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1611678357911/work
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.6.20
cffi @ file:///D:/bld/cffi_1606601414542/work
chardet==3.0.4
click==7.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
comtypes==1.1.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
face-recognition-models==0.3.0
firebase==3.0.1
firebase-admin==4.5.1
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphviz==0.8.4
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.19.0
idna==2.6
importlib-metadata @ file:///D:/bld/importlib-metadata_1610355323632/work
ipykernel @ file:///D:/bld/ipykernel_1611268490473/work/dist/ipykernel-5.4.3-py3-none-any.whl
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1602551949684/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1610375432619/work
jupyter-core @ file:///D:/bld/jupyter_core_1611181880604/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe @ file:///D:/bld/markupsafe_1610127633636/work
matplotlib==3.3.4
mistune @ file:///D:/bld/mistune_1610113131762/work
mkl-service==2.3.0
msgpack==1.0.2
mtcnn==0.1.0
mxnet==1.5.0
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1602859080374/work
nbconvert @ file:///D:/bld/nbconvert_1605402006314/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1611005694671/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1605195931949/work
networkx==2.5
notebook @ file:///D:/bld/notebook_1610575495048/work
numpy @ file:///D:/bld/numpy_1610324703282/work
oauthlib==3.1.0
olefile==0.46
onnx==1.7.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1607785313469/work
pandas @ file:///C:/ci/pandas_1602088210462/work
pandocfilters==1.4.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///C:/ci/pillow_1602785322033/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit==1.0.15
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1610742651773/work
pyparsing==2.4.7
pypiwin32==223
pyrsistent @ file:///D:/bld/pyrsistent_1610146967272/work
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.1
pywin32==300
pywinpty @ file:///D:/bld/pywinpty_1602377668868/work
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7
scikit-learn @ file:///C:/ci/scikit-learn_1598358870446/work
scipy==1.5.4
Send2Trash==1.5.0
simplegeneric==0.8.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
soupsieve==2.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado @ file:///D:/bld/terminado_1609794344119/work
test-generator==0.1.1
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tornado @ file:///D:/bld/tornado_1610094881553/work
traitlets==4.3.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.22
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1603668650351/work
